split sync_union implementation in two different files
